// generated/src/app/api-models/benefit-rule-maint.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class BenefitRuleMaint {
  benefitPackageId: string;
  seqBenPackage: number;
  benefitRule: string;
  type: string;
  operator: string;
  fromValue: string;
  thruValue: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}